services/hunyuan-api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
celery==5.5.3
redis==6.4.0
SQLAlchemy==2.0.43
psycopg[binary]==3.2.9
pydantic-settings==2.10.1
python-multipart==0.0.20
Pillow==11.3.0
boto3==1.39.11
